app/requirements.txt
numpy
pandas
fastapi
uvicorn[standard]
psycopg2-binary
jinja2
python-multipart